Default Vaultwareden to Postgres connection to use sslmode=disable
I've just experienced the error described here:
https://github.com/dani-garcia/vaultwarden/issues/2386

> crypto/evp/evp_enc.c:156: OpenSSL internal error: assertion failed: ctx->cipher->block_size == 1 || ctx->cipher->block_size == 8 || ctx->cipher->block_size == 16

This is with `docker.io/vaultwarden/server:1.26.0-alpine` on `amd64`.

So it seems like it's still a problem, even though the issue is closed.

SSL is not required for our container setup anyway, so we'd better not use it,
which is what this patch does.